Network Tested with 3 ESPs and 1 RPi and works with the strategy topology Wi-Fi Raspberry Pi: Some important corrections

+++ lib/wifi_hal/raspberrypi/wifi_raspberrypi.h
@@ -31,8 +31,8 @@ typedef enum wifi_event_t{
 } wifi_event_t;
 
 typedef struct wifi_event_info__t{
-    int IP[4];
-    int MAC[6];
+    uint8_t IP[4];
+    uint8_t MAC[6];
     int reason;
 }wifi_event_info__t;
